% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summation.R
\name{summation}
\alias{summation}
\title{summation}
\usage{
summation(x, y)
}
\arguments{
\item{x, y}{two values to be summed}
}
\value{
the sum x+y
}
\description{
This is an example of performing a simple summation calculation
}
\examples{
trivial(2, 3)
}
