% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitDData.R
\name{vitDData}
\alias{vitDData}
\title{Table of OV/BV and serum levels}
\format{This table is 675 by 2.\cr}
\source{
??
}
\description{
These values here summarize the OV/BV ratio of 675 patients and the serum
level provided to them.
}
